% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDLY.R
\name{downloadDLY}
\alias{downloadDLY}
\title{Import Canadian Snow Data from .dly file}
\usage{
downloadDLY(fileLoc, progress = FALSE)
}
\arguments{
\item{fileLoc}{File path to .dly data}

\item{progress}{boolean spesifing if you want progress of code to be printed out}
}
\value{
nicely organized dataframe of snow data
}
\description{
Import Canadian Snow Data from .dly file
}
